Updating info function to use engine-api

+++ cluster/cluster.go
@@ -78,7 +78,7 @@ type Cluster interface {
 	TotalMemory() int64
 
 	// Return the number of CPUs in the cluster
-	TotalCpus() int64
+	TotalCpus() int
 
 	// Register an event handler for cluster-wide events.
 	RegisterEventHandler(h EventHandler) error
